requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0
python-dotenv>=1.0.0
requests>=2.32.0
streamlit>=1.39.0
openai>=1.50.0
google-genai
sqlalchemy>=2.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pandas>=2.0.0
